Move changes to SrgDatabase to a subclass
